Refactor all the android code to its own folder
Moved all the jni stuff into a helper class to better isolate the jni
bridge code. Will have to figure out the logging code later.

+++ VPlayerLibrary2/src/main/cpp/player/Player.cpp
@@ -213,7 +213,7 @@ void Player::seek(long positionMill) {
     // Seek to position, if started before duration is gained, bound it later
     long requested = std::max(positionMill, 0L);
     if (mDurationMs) {
-        requested = std::min(requested, mDurationMs);
+        requested = std::min(requested, (long) mDurationMs);
     }
     requested *= MS_TO_TIME_BASE;
